Moved to events.d.ts file
